Send MARSH_CMD_DO_EXPERIMENT_CONTROL when starting/stopping data save
The manager now sends a MARSH_CMD_DO_EXPERIMENT_CONTROL command (with
param1 = 1.0) to all GENERIC and EMG connected nodes, encapsulated
into a COMMAND_LONG message, when the "Save data" button is pressed.

When the "Stop saving" button is pressed, a symmetric
MARSH_CMD_DO_EXPERIMENT_CONTROL command is sent, this time with
param1 = 0.0.

The messages are meant to trigger the start and stop of the acquisition
on remote deviced (specifically, they have been developed with LTW3
RawPower EMG DAQ in mind).